Modify custom exceptions
