+++ src/api/task.py
@@ -1,16 +1,15 @@
 """ Tasks module and APIs"""
 import logging
-from typing import Dict, Any
-
 import requests
-
+from typing import Dict, Any
 from src.constants.constants import SWAN_API, TASK_BIDDING
-from src.exceptions.exceptions import (
-    HTTPError,
-    RequestError,
-    ConnectionError,
-    TimeoutError,
+from src.exceptions.request_exceptions import (
+    SwanRequestError,
+    SwanTimeoutError,
+    SwanConnectionError,
+    SwanHTTPError,
 )
+from src.exceptions.task_exceptions import SwanTaskInvalidInputError
 
 
 def get_task_bidding(task_id) -> Dict[str, Any]:
@@ -35,7 +34,7 @@ def get_task_bidding(task_id) -> Dict[str, Any]:
     """
 
     if not task_id:
-        raise RequestError("Please Provide TASK ID")
+        raise SwanTaskInvalidInputError("Please Provide TASK ID")
 
     base_url = SWAN_API
     # Construct the full URL for the API endpoint
@@ -50,18 +49,18 @@ def get_task_bidding(task_id) -> Dict[str, Any]:
         response.raise_for_status()  # Raises an exception for HTTP errors
     except requests.exceptions.HTTPError as http_err:
         logging.info("HTTP error occurred on TASK_BIDDING API ")
-        raise HTTPError(f"HTTP error occurred: {http_err}")
+        raise SwanHTTPError(f"HTTP error occurred: {http_err}")
 
     except requests.exceptions.ConnectionError as conn_err:
         logging.info("ConnectionError error occurred on TASK_BIDDING API ")
-        raise ConnectionError(f"Connection error occurred: {conn_err}")
+        raise SwanConnectionError(f"Connection error occurred: {conn_err}")
 
     except requests.exceptions.Timeout as timeout_err:
         logging.info("TimeoutError error occurred on TASK_BIDDING API ")
-        raise TimeoutError(f"Timeout error occurred: {timeout_err}")
+        raise SwanTimeoutError(f"Timeout error occurred: {timeout_err}")
 
     except requests.exceptions.RequestException as req_err:
         logging.info("RequestError error occurred on TASK_BIDDING API ")
-        raise RequestError(f"API request failed: {req_err}")
+        raise SwanRequestError(f"API request failed: {req_err}")
 
     return response.json()
