+++ test/tasks/test_task.py
@@ -3,7 +3,7 @@ import pytest
 import requests
 from mock.mock import patch, Mock
 from src.api.task import get_task_bidding
-from src.exceptions.exceptions import RequestError
+from src.exceptions.swan_base_exceptions import RequestError
 
 
 class TestTask:
